UI improved
This is a new UI added. API didn't work and functionality too a little of responsive designed added to the project. This was made on Friday but added today since I wanted to confirm if I was able to modify the project out of the time given on friday. Greetings and thanks for the opportunity. I'm still learning.